[Refactor][Op] Align fused_gdn_gating API and softplus semantics (#10553)
### What this PR does / why we need it?
This PR aligns the AscendC `fused_gdn_gating` operator with the Triton
implementation.

It adds the `softplus_threshold` parameter through the full AscendC op
stack, updates the softplus computation to match Triton semantics, and
supports fp32/bf16/fp16 `A_log` and `dt_bias` inputs while keeping the
actual computation in fp32.

### How was this patch tested?
- Script
``` bash
pytest test/e2e/nightly/single_node/ops/singlecard_ops/test_fused_gdn_gating.py
```
- Result
``` bash
========================== 177 passed in 64.32s (0:01:04) ========================== 
```

- vLLM version: v0.22.1
- vLLM main:
https://github.com/vllm-project/vllm/commit/967c5c3bc38891f4465d3f4e99917ed837bb3833

---------

+++ vllm_ascend/device/device_op.py
@@ -750,7 +750,7 @@ class BaseDeviceAdaptor:
 
     @staticmethod
     def fused_gdn_gating(A_log: torch.Tensor, a: torch.Tensor, b: torch.Tensor, dt_bias: torch.Tensor):
-        return torch.ops._C_ascend.npu_fused_gdn_gating(A_log, a, b, dt_bias.to(torch.float32))
+        return torch.ops._C_ascend.npu_fused_gdn_gating(A_log, a, b, dt_bias.to(A_log.dtype))
 
     @staticmethod
     def split_qkv_rmsnorm_rope(
